requirements.txt
Django==4.2.7
pdfkit==1.0.0
imgkit==1.2.3
pandas==2.0.2
google-api-python-client==2.127.0
google-auth==2.23.3
google-auth-oauthlib==1.2.0